requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
feedparser>=6.0.0
pytz>=2023.3
